pyproject.toml
[project]
name = "blumn"
version = "0.1.0"
description = "Plant care tracking application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "openpyxl>=3.1.2",
    "python-dateutil>=2.8.2",
    "pydantic>=2.6.0",
    "psycopg2>=2.9.10",
    "sqlalchemy>=2.0.42",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 88
target-version = "py313"
